+++ src/components/ActiveAvatar/index.js
