+++ src/components/ClipNewsGridItem/ClipNewsGrid.module.scss
