+++ src/components/NewsGridItem/NewsGridItem.module.scss
