+++ src/containers/HomeChart/index.js
@@ -5,12 +5,12 @@ import { FMP_CLOUD_API_KEY } from "../../apis/apiKeys";
 import { Box } from "@material-ui/core";
 
 const HomeChart = () => {
-  const symbols = ["AAPL", "FB", "MSFT", "AMZN", "GOOG", "NFLX", "SBUX"];
+  const symbols = ["AAPL", "MSFT", "AMZN", "GOOG", "NFLX", "SBUX"];
   const date = new Date();
   const getDate = date.toLocaleDateString();
   const toArrayDate = getDate.split("/");
   const endpoint = (symbol) => {
-    return `https://fmpcloud.io/api/v3/historical-price-full/${symbol}?from=2021-03-14&to=${
+    return `https://fmpcloud.io/api/v3/historical-price-full/${symbol}?from=2021-03-20&to=${
       toArrayDate[2]
     }-${toArrayDate[0]}-${toArrayDate[1] - 1}&apikey=${FMP_CLOUD_API_KEY}`;
   };
