+++ src/containers/HomeChart/index.module.scss
@@ -1,10 +1,11 @@
 .root {
   @media screen and (max-width: 1050px) {
-    display: flex;
-    width: 95vw;
-    overflow-x: auto;
-    background-color: none;
-    background-color: white;
-    margin-top: 8px;
+    // display: flex;
+    // width: 95vw;
+    // overflow-x: auto;
+    // background-color: none;
+    // background-color: white;
+    // margin-top: 8px;
+    display: none;
   }
 }
