+++ src/containers/News/index.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from "react";
 import styles from "./News.module.css";
-import ButtonTab from "../../components/News/ButtonTab";
-import NewsItemLayout from "../../components/News/NewsItemLayout";
+import ButtonTab from "../../components/ButtonTab";
+import NewsItemLayout from "../../components/NewsGridItem";
 import { getNewsData } from "../../apis/rapidApi";
 import { Box } from "@material-ui/core";
 import { Store } from "../../context/Store";
